Add Equipment tests for combat stat getters

+++ lib/model/include/Equipment.h
@@ -21,11 +21,11 @@ namespace model {
         constexpr static int HANDS_ARMOUR_VALUE = 1;
         constexpr static int LEGS_ARMOUR_VALUE = 2;
 
-        constexpr static float WEAPON_CRITICAL_VALUE = 0.02f;
-        constexpr static float BACK_CRITICAL_VALUE= 0.03f;
+        constexpr static double WEAPON_CRITICAL_VALUE = 0.02;
+        constexpr static double BACK_CRITICAL_VALUE = 0.03;
 
-        constexpr static float BACK_DODGE_VALUE = 0.02f;
-        constexpr static float FEET_DODGE_VALUE = 0.03f;
+        constexpr static double BACK_DODGE_VALUE = 0.02;
+        constexpr static double FEET_DODGE_VALUE = 0.03;
 
         /**
          * int: Numbers indicate which slot the item is in
@@ -93,12 +93,24 @@ namespace model {
          */
         bool isEquipmentEmpty();
 
+        /**
+         * @return Returns integer value based on offence modifier of each piece of equipment
+         */
         int getOffenceValue();
 
+        /**
+         * @return Returns integer value based on defence modifier of each piece of equipment
+         */
         int getDefenceValue();
 
+        /**
+         * @return Returns float value based on critical chance modifier of each piece of equipment
+         */
         float getCriticalValue();
 
+        /**
+         * @return Returns float value based on dodge chance modifier of each piece of equipment
+         */
         float getDodgeValue();
     };
 }
